requirements.txt
django==2.2.5
djangorestframework==3.10.3
PyJWT==1.7.1
jwcrypto==0.6.0
python-decouple==3.1
django-cors-headers==3.1.0
boto3==1.9.237
drf-yasg==1.17.0
requests==2.21.0
google-api-python-client==1.7.11
django-haystack==2.8.1
pysolr==3.8.1